feat: rename phoneInput file
